refactor(chat): extract Step-3 dispatch loop to chat_dispatch.py
Move the streaming and non-streaming provider-call-with-failover loops
(chat_completions Step 3, ~570 lines) into src/routes/chat_dispatch.py as
dispatch_streaming() and dispatch_non_streaming(). chat.py drops 1689 -> 1121
lines (-34%); this was the last deeply-coupled piece of the 0d decomposition.

- Request-scoped state is passed as keyword-only args so the moved bodies stay
  verbatim -> zero behavior change.
- The dynamically-injected OpenRouter provider fns and stream_generator are
  resolved through the chat module at call time (_chat.*) so tests patching
  src.routes.chat.* still bind (verified: test_routing_refactor, test_endpoints).
- Removed 23 now-orphaned imports (including 5 pre-existing F401s).
- Documented a pre-existing latent bug: request_model is unbound in the section-4
  trial-tracking call on the authenticated path (always has been); the call has
  always raised-and-been-caught. Behavior preserved; flagged for a separate fix.
- Updated test_pricing_precheck_before_provider_dispatch to follow the moved
  dispatch code (same precedent as the streaming-background test earlier in 0d).

Verified against the full local suite (CI is billing-locked): zero new
regressions vs the clean-HEAD baseline. The only failure-set delta is
xdist-ordering / missing-API-key flakes in untouched provider-client tests,
which pass in isolation.
