+++ src/routes/chat.py
@@ -1,16 +1,12 @@
 import asyncio
 import importlib
-import json
 import logging
-import secrets
 import time
 import uuid
 from contextvars import ContextVar
-from typing import Any
 
-import httpx
 from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Query, Request
-from fastapi.responses import JSONResponse, StreamingResponse
+from fastapi.responses import JSONResponse
 
 import src.db.activity as activity_module
 import src.db.api_keys as api_keys_module
@@ -21,7 +17,7 @@ import src.db.users as users_module
 from src.config import Config
 from src.db.chat_completion_requests_enhanced import save_chat_completion_request_with_cost
 from src.schemas import ProxyRequest
-from src.security.deps import get_api_key, get_optional_api_key
+from src.security.deps import get_optional_api_key
 from src.services.anonymous_rate_limiter import (
     ANONYMOUS_ALLOWED_MODELS,
     ANONYMOUS_DAILY_LIMIT,
@@ -30,17 +26,11 @@ from src.services.anonymous_rate_limiter import (
 )
 from src.services.passive_health_monitor import capture_model_health
 from src.services.prometheus_metrics import (
-    get_trace_exemplar,
-    model_inference_duration,
-    model_inference_requests,
     record_free_model_usage,
-    tokens_used,
 )
-from src.utils.ai_tracing import AIRequestType, AITracer
 from src.utils.exceptions import APIExceptions
 from src.utils.performance_tracker import PerformanceTracker
 from src.utils.rate_limit_headers import get_rate_limit_headers
-from src.utils.sentry_context import capture_provider_error
 
 # Optional Traceloop integration - gracefully handle if not installed
 try:
@@ -51,10 +41,7 @@ except ImportError:
         pass
 
 
-from src.adapters.chat import OpenAIChatAdapter
-
 # Unified chat handler and adapters for chat unification
-from src.handlers.chat_handler import ChatInferenceHandler
 from src.handlers.error_persistence import save_failed_request
 
 # Request correlation ID for distributed tracing
@@ -83,7 +70,6 @@ def braintrust_flush():
 # Import provider registry from canonical module (breaks circular dep with chat_handler.py)
 from src.handlers.provider_registry import (
     PROVIDER_FUNCTIONS,
-    PROVIDER_ROUTING,
     _provider_import_errors,
     _safe_import_provider,
 )
@@ -98,33 +84,17 @@ for _prov_name, _func_names in PROVIDER_FUNCTIONS.items():
         _current_globals[_fn] = _prov_module.get(_fn)
 
 
-def _maybe_record_402(provider: str, status_code: int) -> None:
-    """Record a 402 for provider credit monitoring. Never raises."""
-    if status_code == 402:
-        try:
-            from src.services.provider_credit_monitor import record_provider_402
-
-            record_provider_402(provider)
-        except Exception:
-            pass
-
 import src.services.rate_limiting as rate_limiting_service
 import src.services.trial_validation as trial_module
 
-from src.services.model_transformations import transform_model_id
+# Step-3 dispatch (streaming / non-streaming provider-call loops) lives in chat_dispatch.
+from src.routes.chat_dispatch import dispatch_non_streaming, dispatch_streaming  # noqa: E402
 from src.security.inference_gates import (
     enforce_anonymous_gate,
     enforce_model_pricing_gate,
     enforce_subscription_status_gate,
 )
-from src.services.pricing import calculate_cost_async, get_model_pricing_async, model_has_pricing
-from src.services.provider_failover import (
-    build_provider_failover_chain,
-    enforce_model_failover_rules,
-    filter_by_circuit_breaker,
-    map_provider_error,
-    should_failover,
-)
+from src.services.pricing import calculate_cost_async, get_model_pricing_async
 from src.utils.token_estimator import estimate_message_tokens
 
 
@@ -200,11 +170,6 @@ router = APIRouter()
 logger.info("🔄 Chat module initialized - router created")
 logger.info(f"   Router type: {type(router)}")
 
-DEFAULT_PROVIDER_TIMEOUT = 30
-PROVIDER_TIMEOUTS = {
-    "huggingface": 120,
-    "near": 120,  # Large models like Qwen3-30B need extended timeout
-}
 
 # Auto-routing constants
 # Using "router" prefix to avoid confusion with OpenRouter's "openrouter/auto" model
@@ -345,10 +310,9 @@ def _fallback_get_user(api_key: str):
 
 
 from src.routes.chat_context import inject_conversation_history, persist_conversation_turn
-from src.routes.chat_routing import resolve_model_routing
 from src.routes.chat_request import prepare_upstream_request
+from src.routes.chat_routing import resolve_model_routing
 from src.routes.chat_streaming import stream_generator  # noqa: F401
-
 
 # Log route registration for debugging
 logger.info("📍 Registering /chat/completions endpoint")
@@ -400,7 +364,6 @@ async def chat_completions(
         request_id=request_id,
         api_key_mask=mask_key(api_key) if api_key else "anonymous",
     )
-
 
     # Initialize performance tracker
     tracker = PerformanceTracker(endpoint="/v1/chat/completions")
@@ -845,579 +808,43 @@ async def chat_completions(
 
         # === 3) Call upstream (streaming or non-streaming) ===
         if req.stream:
-            # Streaming path
-            # Use unified handler for authenticated streaming requests
-            if not is_anonymous:
-                # Authenticated streaming with provider failover.
-                # We wrap consumption inside an async generator so that errors during
-                # iteration (not just setup) can be caught and the next provider tried,
-                # but only before the first content chunk has been sent to the client.
-                async def _auth_stream_with_failover():
-                    _adapter = OpenAIChatAdapter()
-                    last_exc = None
-
-                    for _idx, _attempt_provider in enumerate(provider_chain):
-                        _is_last = _idx == len(provider_chain) - 1
-                        _content_started = False
-                        try:
-                            _internal_req = _adapter.to_internal_request(
-                                {
-                                    "messages": messages,
-                                    "model": original_model,
-                                    "stream": True,
-                                    **optional,
-                                }
-                            )
-                            # Set the provider for this attempt
-                            _internal_req.provider = _attempt_provider
-
-                            _handler = ChatInferenceHandler(
-                                api_key, background_tasks, request=request
-                            )
-                            _internal_stream = _handler.process_stream(_internal_req)
-                            _sse_stream = _adapter.from_internal_stream(_internal_stream)
-
-                            async for _chunk in _sse_stream:
-                                _content_started = True
-                                yield _chunk
-
-                            return  # Stream completed successfully
-
-                        except HTTPException as _http_exc:
-                            if not _content_started and not _is_last and should_failover(_http_exc):
-                                logger.warning(
-                                    f"[Unified Handler] Provider '{_attempt_provider}' failed "
-                                    f"(HTTP {_http_exc.status_code}) for model {original_model}, "
-                                    f"failing over ({_idx + 1}/{len(provider_chain)})"
-                                )
-                                last_exc = _http_exc
-                                continue
-                            raise
-
-                        except Exception as _exc:
-                            if not _content_started and not _is_last:
-                                logger.warning(
-                                    f"[Unified Handler] Provider '{_attempt_provider}' error "
-                                    f"({type(_exc).__name__}: {_exc}) for model {original_model}, "
-                                    f"failing over ({_idx + 1}/{len(provider_chain)})"
-                                )
-                                last_exc = map_provider_error(
-                                    _attempt_provider, original_model, _exc
-                                )
-                                continue
-                            if isinstance(_exc, HTTPException):
-                                raise
-                            raise map_provider_error(
-                                _attempt_provider, original_model, _exc
-                            ) from _exc
-
-                    # All providers exhausted
-                    if last_exc:
-                        raise last_exc
-
-                # Prepare response headers
-                stream_headers = {}
-                if rl_pre is not None:
-                    stream_headers.update(get_rate_limit_headers(rl_pre))
-                if tracker:
-                    prep_time_ms = tracker.get_total_duration() * 1000
-                    stream_headers["X-Prep-Time-Ms"] = f"{prep_time_ms:.1f}"
-                stream_headers["X-Provider"] = "unified"
-                stream_headers["X-Model"] = original_model
-                stream_headers["X-Requested-Model"] = original_model
-                # SSE streaming headers to prevent buffering by proxies/nginx
-                stream_headers["X-Accel-Buffering"] = "no"
-                stream_headers["Cache-Control"] = "no-cache, no-transform"
-                stream_headers["Connection"] = "keep-alive"
-
-                logger.info(
-                    f"[Unified Handler] Returning SSE streaming response for model {original_model} "
-                    f"(provider_chain={provider_chain})"
-                )
-
-                return StreamingResponse(
-                    _auth_stream_with_failover(),
-                    media_type="text/event-stream",
-                    headers=stream_headers,
-                )
-            else:
-                # Anonymous users: keep existing provider routing logic
-                last_http_exc = None
-                for idx, attempt_provider in enumerate(provider_chain):
-                    attempt_model = transform_model_id(original_model, attempt_provider)
-                    if attempt_model != original_model:
-                        logger.info(
-                            f"Transformed model ID from '{original_model}' to '{attempt_model}' for provider {attempt_provider}"
-                        )
-
-                    request_model = attempt_model
-                    is_async_stream = False  # Default to sync, only OpenRouter uses async currently
-                    try:
-                        # Registry-based provider dispatch (replaces ~400 lines of if-elif chains)
-                        # Note: Streaming tracing is handled in stream_generator to capture final token counts
-                        if attempt_provider == "fal":
-                            # FAL models are for image/video generation, not chat completions
-                            raise HTTPException(
-                                status_code=400,
-                                detail={
-                                    "error": {
-                                        "message": f"Model '{request_model}' is a FAL.ai image/video generation model "
-                                        "and is not available through the chat completions endpoint. "
-                                        "Please use the /v1/images/generations endpoint with provider='fal' instead.",
-                                        "type": "invalid_request_error",
-                                        "code": "model_not_supported_for_chat",
-                                    }
-                                },
-                            )
-                        elif attempt_provider in PROVIDER_ROUTING:
-                            # Use registry for all registered providers
-                            stream_func = PROVIDER_ROUTING[attempt_provider]["stream"]
-                            stream = await _to_thread(
-                                stream_func, messages, request_model, **optional
-                            )
-                        else:
-                            # Default to OpenRouter with async streaming for performance
-                            try:
-                                stream = (
-                                    await make_openrouter_request_openai_stream_async(  # noqa: F821
-                                        messages, request_model, **optional
-                                    )
-                                )
-                                is_async_stream = True
-                                logger.debug(
-                                    f"Using async streaming for OpenRouter model {request_model}"
-                                )
-                            except Exception as async_err:
-                                # Fallback to sync streaming if async fails
-                                logger.warning(
-                                    f"Async streaming failed, falling back to sync: {async_err}"
-                                )
-                                stream = await _to_thread(
-                                    make_openrouter_request_openai_stream,  # noqa: F821
-                                    messages,
-                                    request_model,
-                                    **optional,
-                                )
-                                is_async_stream = False
-
-                        provider = attempt_provider
-                        model = request_model
-                        # Get rate limit headers if available (pre-stream check)
-                        stream_headers = {}
-                        if rl_pre is not None:
-                            stream_headers.update(get_rate_limit_headers(rl_pre))
-
-                        # PERF: Add timing headers for debugging stream startup latency
-                        if tracker:
-                            prep_time_ms = tracker.get_total_duration() * 1000
-                            stream_headers["X-Prep-Time-Ms"] = f"{prep_time_ms:.1f}"
-                        stream_headers["X-Provider"] = provider
-                        stream_headers["X-Model"] = model
-                        stream_headers["X-Requested-Model"] = original_model
-
-                        # SSE streaming headers to prevent buffering by proxies/nginx
-                        stream_headers["X-Accel-Buffering"] = "no"
-                        stream_headers["Cache-Control"] = "no-cache, no-transform"
-                        stream_headers["Connection"] = "keep-alive"
-
-                        return StreamingResponse(
-                            stream_generator(
-                                stream,
-                                user,
-                                api_key,
-                                model,
-                                trial,
-                                environment_tag,
-                                session_id,
-                                messages,
-                                rate_limit_mgr,
-                                provider,
-                                tracker,
-                                is_anonymous,
-                                is_async_stream=is_async_stream,
-                                request_id=request_id,
-                                api_key_id=api_key_id,
-                                client_ip=client_ip if is_anonymous else None,
-                                request=request,
-                            ),
-                            media_type="text/event-stream",
-                            headers=stream_headers,
-                        )
-                    except Exception as exc:
-                        if isinstance(exc, httpx.TimeoutException | asyncio.TimeoutError):
-                            logger.warning("Upstream timeout (%s): %s", attempt_provider, exc)
-                            # Capture timeout to Sentry
-                            capture_provider_error(
-                                exc,
-                                provider=attempt_provider,
-                                model=request_model,
-                                endpoint="/v1/chat/completions",
-                                request_id=request_id_var.get(),
-                            )
-                        elif isinstance(exc, httpx.RequestError):
-                            logger.warning("Upstream network error (%s): %s", attempt_provider, exc)
-                            # Capture network error to Sentry
-                            capture_provider_error(
-                                exc,
-                                provider=attempt_provider,
-                                model=request_model,
-                                endpoint="/v1/chat/completions",
-                                request_id=request_id_var.get(),
-                            )
-                        elif isinstance(exc, httpx.HTTPStatusError):
-                            logger.debug(
-                                "Upstream HTTP error (%s): %s",
-                                attempt_provider,
-                                exc.response.status_code,
-                            )
-                            # Capture HTTP errors to Sentry (except 4xx client errors)
-                            if exc.response.status_code >= 500:
-                                capture_provider_error(
-                                    exc,
-                                    provider=attempt_provider,
-                                    model=request_model,
-                                    endpoint="/v1/chat/completions",
-                                    request_id=request_id_var.get(),
-                                )
-                        else:
-                            logger.error(
-                                "Unexpected upstream error (%s): %s", attempt_provider, exc
-                            )
-                            # Capture unexpected errors to Sentry
-                            capture_provider_error(
-                                exc,
-                                provider=attempt_provider,
-                                model=request_model,
-                                endpoint="/v1/chat/completions",
-                                request_id=request_id_var.get(),
-                            )
-                        http_exc = map_provider_error(attempt_provider, request_model, exc)
-
-                        last_http_exc = http_exc
-
-                        # Record 402 for provider credit monitoring BEFORE failover
-                        # (must run before should_failover/continue skips past it)
-                        _maybe_record_402(attempt_provider, http_exc.status_code)
-
-                        if idx < len(provider_chain) - 1 and should_failover(http_exc):
-                            next_provider = provider_chain[idx + 1]
-                            logger.warning(
-                                "Provider '%s' failed with status %s (%s). Falling back to '%s'.",
-                                attempt_provider,
-                                http_exc.status_code,
-                                http_exc.detail,
-                                next_provider,
-                            )
-                            continue
-
-                        # If this is a 402 (Payment Required) error and we've exhausted the chain,
-                        # rebuild the chain allowing payment failover to try alternative providers
-                        if http_exc.status_code == 402 and idx == len(provider_chain) - 1:
-                            extended_chain = build_provider_failover_chain(provider)
-                            extended_chain = enforce_model_failover_rules(
-                                original_model, extended_chain, allow_payment_failover=True
-                            )
-                            extended_chain = filter_by_circuit_breaker(
-                                original_model, extended_chain
-                            )
-                            # Find providers we haven't tried yet
-                            new_providers = [p for p in extended_chain if p not in provider_chain]
-                            if new_providers:
-                                logger.warning(
-                                    "Provider '%s' returned 402 (Payment Required). "
-                                    "Extending failover chain with: %s",
-                                    attempt_provider,
-                                    new_providers,
-                                )
-                                provider_chain.extend(new_providers)
-                                continue
-
-                        raise http_exc
-
-                raise last_http_exc or HTTPException(status_code=502, detail="Upstream error")
+            return await dispatch_streaming(
+                is_anonymous=is_anonymous,
+                provider_chain=provider_chain,
+                messages=messages,
+                original_model=original_model,
+                optional=optional,
+                api_key=api_key,
+                api_key_id=api_key_id,
+                background_tasks=background_tasks,
+                request=request,
+                request_id=request_id,
+                rl_pre=rl_pre,
+                tracker=tracker,
+                user=user,
+                trial=trial,
+                environment_tag=environment_tag,
+                session_id=session_id,
+                rate_limit_mgr=rate_limit_mgr,
+                client_ip=client_ip,
+            )
 
         # Non-streaming response
         start = time.monotonic()
-        processed = None
-        last_http_exc = None
-
-        # Use unified handler for authenticated non-streaming requests
-        if not is_anonymous:
-            try:
-                logger.info(
-                    f"[Unified Handler] Processing authenticated non-streaming request for model {original_model}"
-                )
-
-                # Convert external OpenAI format to internal format
-                adapter = OpenAIChatAdapter()
-                internal_request = adapter.to_internal_request(
-                    {"messages": messages, "model": original_model, "stream": False, **optional}
-                )
-                # Pass the detected provider so the handler doesn't have to re-detect
-                internal_request.provider = provider
-
-                # Create unified handler with user context (pass request for disconnect detection)
-                handler = ChatInferenceHandler(api_key, background_tasks, request=request)
-
-                # Wrap with AITracer for gen_ai.* telemetry + track duration for Prometheus
-                inference_start = time.time()
-                async with AITracer.trace_inference(
-                    provider="onerouter",  # Will be updated after response
-                    model=original_model,
-                    request_type=AIRequestType.CHAT_COMPLETION,
-                    operation_name=f"unified_handler/{original_model}",
-                ) as trace_ctx:
-                    # Process request through unified pipeline
-                    internal_response = await handler.process(internal_request)
-
-                    # Convert internal response back to OpenAI format
-                    processed = adapter.from_internal_response(internal_response)
-
-                    # Extract values for postprocessing
-                    provider = internal_response.provider_used or "onerouter"
-                    model = internal_response.model or original_model
-
-                    # Set trace attributes with actual values from response
-                    usage = processed.get("usage", {}) or {}
-                    input_tokens = usage.get("prompt_tokens", 0)
-                    output_tokens = usage.get("completion_tokens", 0)
-                    trace_ctx.set_token_usage(
-                        input_tokens=input_tokens,
-                        output_tokens=output_tokens,
-                        total_tokens=usage.get("total_tokens", 0),
-                    )
-                    trace_ctx.set_response_model(
-                        response_model=model,
-                        finish_reason=(
-                            processed.get("choices", [{}])[0].get("finish_reason")
-                            if processed.get("choices")
-                            else None
-                        ),
-                        response_id=processed.get("id"),
-                    )
-                    if optional:
-                        trace_ctx.set_model_parameters(
-                            temperature=optional.get("temperature"),
-                            max_tokens=optional.get("max_tokens"),
-                            top_p=optional.get("top_p"),
-                            frequency_penalty=optional.get("frequency_penalty"),
-                            presence_penalty=optional.get("presence_penalty"),
-                        )
-                    if user:
-                        trace_ctx.set_user_info(
-                            user_id=str(user.get("id")),
-                            tier="trial" if trial.get("is_trial") else "paid",
-                        )
-
-                # Record Prometheus metrics for model popularity tracking
-                inference_duration = time.time() - inference_start
-                exemplar = get_trace_exemplar()
-                model_inference_requests.labels(
-                    provider=provider, model=model, status="success"
-                ).inc(1, exemplar=exemplar)
-                model_inference_duration.labels(provider=provider, model=model).observe(
-                    inference_duration, exemplar=exemplar
-                )
-                if input_tokens > 0:
-                    tokens_used.labels(provider=provider, model=model, token_type="input").inc(
-                        input_tokens, exemplar=exemplar
-                    )
-                if output_tokens > 0:
-                    tokens_used.labels(provider=provider, model=model, token_type="output").inc(
-                        output_tokens, exemplar=exemplar
-                    )
-
-                logger.info(
-                    f"[Unified Handler] Successfully processed request: provider={provider}, model={model}"
-                )
-
-            except Exception as exc:
-                # Record error metric for model popularity tracking
-                error_provider = provider if "provider" in locals() else "onerouter"
-                error_model = model if "model" in locals() else original_model
-                model_inference_requests.labels(
-                    provider=error_provider, model=error_model, status="error"
-                ).inc(1, exemplar=get_trace_exemplar())
-
-                # Map any errors to HTTPException
-                logger.error(f"[Unified Handler] Error: {type(exc).__name__}: {exc}", exc_info=True)
-                if isinstance(exc, HTTPException):
-                    raise
-                # Map provider-specific errors (map_provider_error imported at module level)
-                http_exc = map_provider_error(error_provider, error_model, exc)
-                raise http_exc
-        else:
-            # Anonymous users: keep existing provider routing logic
-            for idx, attempt_provider in enumerate(provider_chain):
-                attempt_model = transform_model_id(original_model, attempt_provider)
-                if attempt_model != original_model:
-                    logger.info(
-                        f"Transformed model ID from '{original_model}' to '{attempt_model}' for provider {attempt_provider}"
-                    )
-
-                request_model = attempt_model
-                request_timeout = PROVIDER_TIMEOUTS.get(attempt_provider, DEFAULT_PROVIDER_TIMEOUT)
-                if request_timeout != DEFAULT_PROVIDER_TIMEOUT:
-                    logger.debug(
-                        "Using extended timeout %ss for provider %s",
-                        request_timeout,
-                        attempt_provider,
-                    )
-
-                try:
-                    # Registry-based provider dispatch (replaces ~400 lines of if-elif chains)
-                    # Wrap provider calls with distributed tracing for Tempo
-                    async with AITracer.trace_inference(
-                        provider=attempt_provider,
-                        model=request_model,
-                        request_type=AIRequestType.CHAT_COMPLETION,
-                    ) as trace_ctx:
-                        if attempt_provider == "fal":
-                            # FAL models are for image/video generation, not chat completions
-                            raise HTTPException(
-                                status_code=400,
-                                detail={
-                                    "error": {
-                                        "message": f"Model '{request_model}' is a FAL.ai image/video generation model "
-                                        "and is not available through the chat completions endpoint. "
-                                        "Please use the /v1/images/generations endpoint with provider='fal' instead.",
-                                        "type": "invalid_request_error",
-                                        "code": "model_not_supported_for_chat",
-                                    }
-                                },
-                            )
-                        elif attempt_provider in PROVIDER_ROUTING:
-                            # Use registry for all registered providers
-                            request_func = PROVIDER_ROUTING[attempt_provider]["request"]
-                            process_func = PROVIDER_ROUTING[attempt_provider]["process"]
-                            resp_raw = await asyncio.wait_for(
-                                _to_thread(request_func, messages, request_model, **optional),
-                                timeout=request_timeout,
-                            )
-                            processed = await _to_thread(process_func, resp_raw)
-                        else:
-                            # Default to OpenRouter
-                            resp_raw = await asyncio.wait_for(
-                                _to_thread(
-                                    make_openrouter_request_openai,  # noqa: F821
-                                    messages,
-                                    request_model,
-                                    **optional,
-                                ),
-                                timeout=request_timeout,
-                            )
-                            processed = await _to_thread(
-                                process_openrouter_response,  # noqa: F821
-                                resp_raw,
-                            )
-
-                        # Extract token usage from response for tracing
-                        usage = processed.get("usage", {}) or {}
-                        trace_prompt_tokens = usage.get("prompt_tokens", 0)
-                        trace_completion_tokens = usage.get("completion_tokens", 0)
-                        trace_total_tokens = usage.get("total_tokens", 0)
-
-                        # Calculate cost for tracing
-                        trace_cost = await calculate_cost_async(
-                            request_model, trace_prompt_tokens, trace_completion_tokens
-                        )
-
-                        # Set token usage and cost on trace span
-                        trace_ctx.set_token_usage(
-                            input_tokens=trace_prompt_tokens,
-                            output_tokens=trace_completion_tokens,
-                            total_tokens=trace_total_tokens,
-                        )
-                        trace_ctx.set_cost(trace_cost)
-
-                        # Set actual response model (may differ from requested model)
-                        response_model = processed.get("model", request_model)
-                        # Extract finish reason from first choice if available
-                        choices = processed.get("choices", [])
-                        finish_reason = choices[0].get("finish_reason") if choices else None
-                        response_id = processed.get("id")
-                        trace_ctx.set_response_model(
-                            response_model=response_model,
-                            finish_reason=finish_reason,
-                            response_id=response_id,
-                        )
-
-                        # Set model parameters if available
-                        if optional:
-                            trace_ctx.set_model_parameters(
-                                temperature=optional.get("temperature"),
-                                max_tokens=optional.get("max_tokens"),
-                                top_p=optional.get("top_p"),
-                                frequency_penalty=optional.get("frequency_penalty"),
-                                presence_penalty=optional.get("presence_penalty"),
-                            )
-
-                        # Set user info if authenticated
-                        if not is_anonymous and user:
-                            trace_ctx.set_user_info(
-                                user_id=str(user.get("id")),
-                                tier="trial" if trial.get("is_trial") else "paid",
-                            )
-
-                    provider = attempt_provider
-                    model = request_model
-                    break
-                except Exception as exc:
-                    if isinstance(exc, httpx.TimeoutException | asyncio.TimeoutError):
-                        logger.warning("Upstream timeout (%s): %s", attempt_provider, exc)
-                    elif isinstance(exc, httpx.RequestError):
-                        logger.warning("Upstream network error (%s): %s", attempt_provider, exc)
-                    elif isinstance(exc, httpx.HTTPStatusError):
-                        logger.debug(
-                            "Upstream HTTP error (%s): %s",
-                            attempt_provider,
-                            exc.response.status_code,
-                        )
-                    else:
-                        logger.error("Unexpected upstream error (%s): %s", attempt_provider, exc)
-                    http_exc = map_provider_error(attempt_provider, request_model, exc)
-
-                    last_http_exc = http_exc
-
-                    # Record 402 for provider credit monitoring BEFORE failover
-                    _maybe_record_402(attempt_provider, http_exc.status_code)
-
-                    if idx < len(provider_chain) - 1 and should_failover(http_exc):
-                        next_provider = provider_chain[idx + 1]
-                        logger.warning(
-                            "Provider '%s' failed with status %s (%s). Falling back to '%s'.",
-                            attempt_provider,
-                            http_exc.status_code,
-                            http_exc.detail,
-                            next_provider,
-                        )
-                        continue
-
-                    # If this is a 402 (Payment Required) error and we've exhausted the chain,
-                    # rebuild the chain allowing payment failover to try alternative providers
-                    if http_exc.status_code == 402 and idx == len(provider_chain) - 1:
-                        extended_chain = build_provider_failover_chain(provider)
-                        extended_chain = enforce_model_failover_rules(
-                            original_model, extended_chain, allow_payment_failover=True
-                        )
-                        extended_chain = filter_by_circuit_breaker(original_model, extended_chain)
-                        # Find providers we haven't tried yet
-                        new_providers = [p for p in extended_chain if p not in provider_chain]
-                        if new_providers:
-                            logger.warning(
-                                "Provider '%s' returned 402 (Payment Required). "
-                                "Extending failover chain with: %s",
-                                attempt_provider,
-                                new_providers,
-                            )
-                            provider_chain.extend(new_providers)
-                            continue
-
-                    raise http_exc
-
-        if processed is None:
-            raise last_http_exc or HTTPException(status_code=502, detail="Upstream error")
-
+        processed, provider, model = await dispatch_non_streaming(
+            is_anonymous=is_anonymous,
+            provider_chain=provider_chain,
+            messages=messages,
+            original_model=original_model,
+            optional=optional,
+            model=model,
+            provider=provider,
+            api_key=api_key,
+            background_tasks=background_tasks,
+            request=request,
+            user=user,
+            trial=trial,
+        )
         elapsed = max(0.001, time.monotonic() - start)
 
         # === 4) Usage, pricing, final checks ===
@@ -1439,12 +866,18 @@ async def chat_completions(
 
             if trial.get("is_trial") and not trial.get("is_expired"):
                 try:
+                    # NOTE: `request_model` was a dispatch-loop local (set only on the
+                    # anonymous branch) and is unbound on this authenticated path — it was
+                    # already unbound here before the Step-3 dispatch extraction, so this
+                    # call has always raised and been caught (trial usage is not tracked
+                    # here). Behavior preserved verbatim; fixing it is out of scope for this
+                    # refactor (would change trial accounting). Tracked separately.
                     await _to_thread(
                         track_trial_usage,
                         api_key,
                         total_tokens,
                         1,
-                        model_id=request_model,
+                        model_id=request_model,  # noqa: F821  (pre-existing unbound; see note above)
                         prompt_tokens=prompt_tokens,
                         completion_tokens=completion_tokens,
                     )
@@ -1673,7 +1106,6 @@ async def chat_completions(
         raise HTTPException(
             status_code=500, detail=f"Internal server error (request ID: {request_id})"
         )
-
 
 
 # Log successful module load - this should appear in startup logs if chat.py loads correctly
